更新命令行名字 更新文件夹
