--- lib/scss-lib/import-lib-settings.scss
@@ -1,14 +1,0 @@
-//页面基准大小
-$pageWidth: 320px !default;
-$baseFontSize: 20px !default;
-
-//常用间距 xs s xm m xl l
-$spaces: 5 10 15 20 !default;
-
-//常用字体颜色
-$textColors: f60 111 333 666 888 999 fff ccc 369 !default;
-
-//圆角
-$smallRadius: 3px;
-$basicRadius: 5px;
-$bigRadius: 10px;
