--- lib/scss-lib/import-mod-gap-block-rem.scss
@@ -1,41 +1,0 @@
-.mod-comm-gap-block {
-  background-color: #fff;
-  border-top: 1px solid #e5e5e5;
-  border-bottom: 1px solid #e5e5e5;
-  padding-left: pxToRem(15);
-  //@include onepx(top bottom, #e2e2e2, 30%, darken);
-  .line-item {
-    position: relative;
-    width: 100%;
-    display: -webkit-box;
-    -webkit-box-pack:justify;
-    -webkit-box-align: center;
-    border-bottom: 1px solid #e5e5e5;
-    //@include onepx(bottom, #e2e2e2, 30%, darken);
-    padding: pxToRem(10) 0;
-    box-sizing: border-box;
-
-    &:last-child {
-      border-bottom: none;
-      //@include onepx(bottom, transparent);
-    }
-
-    label, .desc-left {
-      display: block;
-      width: 25%;
-    }
-
-    .desc-right {
-      -webkit-box-flex: 1;
-      padding-right: pxToRem(15);
-    }
-  }
-
-  &.border-full {
-    padding-left: 0;
-    padding-right: 0;
-    .desc-left {
-      padding-left: pxToRem(15);
-    }
-  }
-}
