--- lib/scss-lib/import-mod-switch-rem.scss
@@ -1,27 +1,0 @@
-.mod-comm-switch-btn {
-  background-color: #ff6600;
-  width: 2.6rem;
-  height: 1.4rem;
-  border-radius: .8rem;
-  position: absolute;
-  right: pxToRem(15);
-  top: 50%;
-  margin-top: -.7rem;
-  &:after {
-    background-color: #fff;
-    @include cycle-lib(1.3rem, 1.3rem, 999px);
-    position: absolute;
-    right: 1px;
-    top: 50%;
-    margin-top: -.65rem;
-    -webkit-transition: right 0.15s linear;
-    content:"";
-  }
-  &.close {
-    background-color: #ccc;
-    &:after {
-      right: 1.22rem;
-      -webkit-transition: right 0.15s linear;
-    }
-  }
-}
