+++ package.json
@@ -1,16 +1,20 @@
 {
-  "name": "jyb-command-line",
-  "version": "1.0.0",
+  "name": "easy-workflow",
+  "version": "0.0.1",
   "private": true,
-  "description": "加油宝前端工作流命令行工具",
-  "licenses": [
-    {
-      "type": "MIT"
-    }
-  ],
-  "author": ["kardelchen","haiyang"],
+  "description": "前端构建工作流命",
+  "keywords":["workflow","css", "static", "node", "gulp"],
+  "bugs": {
+      "email": "kardelchen@jyblife.com"
+  },
+  "licenses": "MIT",
+  "contributors": ["kardelchen","haiyang"],
+  "repository": {
+      "type": "git",
+      "url": "https://github.com/grassmu/easy-workflow.git"
+  },
   "bin": {
-    "jyb": "./bin/command"
+    "es": "./bin/command"
   },
   "dependencies": {
     "colors": "^1.1.2",
