+++ templates/package.json
