+++ templates/scss-lib/import-mod-amount-input-rem.scss
@@ -1,0 +1,86 @@
+.mod-comm-buy-amount-wrap {
+  width: pxToRem(151);
+  border-radius: 3px;
+  height: pxToRem(37);
+  position: relative;
+  text-align: center;
+  float: right;
+  .buy-mp {
+    background-color: #ff6600;
+    width: pxToRem(37);
+    text-align: center;
+    color: #999;
+    height: 100%;
+    @include vt-middle();
+    position: absolute;
+    top: 0;
+    @include user-select();
+    &.buy-mp_minus {
+      border-top-left-radius: 5px;
+      border-bottom-left-radius: 5px;
+      left: 1px;
+      text-align: center;
+      .minus-icon {
+        height: pxToRem(4);
+        width: pxToRem(18);
+        background-color: #fff;
+        display: inline-block;
+        vertical-align: middle;
+        position: relative;
+        top: -.1rem;
+      }
+    }
+    &.buy-mp_plus {
+      border-top-right-radius: 3px;
+      border-bottom-right-radius: 3px;
+      right: 1px;
+      .plus-icon {
+        height: pxToRem(4);
+        width: pxToRem(18);
+        background-color: #fff;
+        display: inline-block;
+        vertical-align: middle;
+        position: relative;
+        top: -.1rem;
+        &:after {
+          content: "";
+          -webkit-transform: rotate(90deg);
+          height: pxToRem(4);
+          width: pxToRem(18);
+          position: absolute;
+          top: 0;
+          left:0;
+          background-color: #fff;
+        }
+      }
+    }
+    &.buy-mp_disabled {
+      background-color: #e5e5e5;
+      .minus-icon {
+        background-color: #cccccc;
+      }
+      .plus-icon {
+        background-color: #cccccc;
+        &:after {
+          background-color: #cccccc;
+        }
+      }
+    }
+  }
+  .buy-input {
+    width: 3.8rem;
+    height: 1.85rem;
+    font-size: .85rem;
+    text-align: center;
+    outline: none;
+    -webkit-appearance: none;
+    border-top: 1px solid #e5e5e5;
+    border-bottom:1px solid #e5e5e5;
+    border-left: 0;
+    border-right: 0;
+    -webkit-border-radius: 0;
+    -moz-border-radius: 0;
+    border-radius: 0;
+    -webkit-box-sizing: border-box;
+  }
+}
